prevent socket buffer deadlock
A potential socket buffer deadlock can occur when using ActiveMQ
Classic. The issue manifests as Net::Stomp being blocked on a socket
write, while the corresponding thread in ActiveMQ is also blocked on a
socket write. This situation presumably arises because the read and
write buffers on both sides of the connection are full.

The deadlock can be triggered by having a queue consumer spamming that
queue with messages that get increasingly larger.

I have not been able to reproduce the issue against ActiveMQ Artemis.

To prevent this issue, we switch to using a non-blocking socket. If syswrite
returns EWOULDBLOCK, we attempt to read one message from the socket (any
received message is buffered) before retrying the syswrite operation. This
ensures communication can continue without becoming stuck.